Search for child selectors, don't show duplicates.
Whereas before, if an element looked like the following:
`#an-element-here #a-child-element`, it would get passed
through just like that, we now split each one so it's
like this:

[
    #an-element-here,
    #a-child-element
]

This means that `mort` will look for both elements instead
of blindly skipping child selectors.

We have also removed duplicate selectors from appearing in the
output (since it's pointless).

Finally, we have added the "lib: ["es7"]" option to tsconfig
so we can use the `includes` function on arrays.